ex01

+++ ex01/Span.cpp
@@ -6,7 +6,7 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/25 13:12:14 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2026/02/25 15:01:24 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2026/02/25 15:24:34 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,30 +38,30 @@ void Span::addNumber(int val) {
     _v.push_back(val);
 }
 
-int Span::shortestSpan() const {
+long int Span::shortestSpan() const {
     std::vector<int> copy;
-    int min;
+    long min;
     size_t size;
 
     if (this->_v.size() <= 1)
         throw NoSpanException();
     copy = this->_v;
     std::sort(copy.begin(), copy.end());
-    min = copy[1] - copy[0];
+    min = static_cast<long int>(copy[1]) - static_cast<long int>(copy[0]);
     size = copy.size();
     for (size_t i = 1; i + 1 < size; i++) {
-        if (copy[i + 1] - copy[i] < min)
-            min = copy[i + 1] - copy[i];
+        if (static_cast<long int>(copy[i + 1]) - static_cast<long int>(copy[i]) < min)
+            min = static_cast<long int>(copy[i + 1]) - static_cast<long int>(copy[i]);
     }
     return (min);
 }
 
-int Span::longestSpan() const {
+long int Span::longestSpan() const {
     std::vector<int> copy;
 
     if (this->_v.size() <= 1)
         throw NoSpanException();
     copy = this->_v;
     std::sort(copy.begin(), copy.end());
-    return (copy[copy.size() - 1] - copy[0]);
+    return (static_cast<long int>(copy[copy.size() - 1]) - static_cast<long int>(copy[0]));
 }
